+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - `Heading` component
 - `Paragraph` component
+- `SummaryList` component
 
 ## [0.0.1] - 26-11-2019